Refactor runtime to make entrypoint explicit
The entrypoint was previously in a submodule and had a somewhat misleading name.
